code formating added

+++ statistics-repository/src/main/java/com/banadiga/acontrol/statistics/module/gwt/One.java
@@ -10,6 +10,6 @@ import lombok.Data;
 public class One implements Serializable {
   private String startTime;
   private String endTime;
-  private int amount;
+  private long amount;
 
 }
